refactor(traces): evict calculateAggregatedUsage from the tooltip component (LFE-14804)
A component file exports the component and its types, nothing else (rule 2).
calculateAggregatedUsage was a 40-line pure reducer living in
BreakdownTooltip.tsx — and it was already being imported from there by
src/components/table/use-cases/observations.tsx, so it was public API smuggled
out of a component.

Now fns/calculateAggregatedUsage.ts, exposed through the feature surface, and
the tooltip imports only the `Details` type it shares with it.

rule 2: 3 -> 2.
