+++ web/src/features/traces/index.ts
@@ -16,6 +16,7 @@ export { traceDetailTitle } from "@/src/features/traces/fns/traceDetailTitle";
 export { useTraceDetailData } from "@/src/features/traces/hooks/useTraceDetailData";
 
 export { BreakdownTooltip } from "@/src/features/traces/components/BreakdownTooltip";
+export { calculateAggregatedUsage } from "@/src/features/traces/fns/calculateAggregatedUsage";
 export { CopyIdsPopover } from "@/src/features/traces/components/CopyIdsPopover";
 
 export {
